fw-6473, extract permission tests and async test base
- and split tests to have one api per file
